Fixed pitch at 90 degrees and imgui tree id bug

+++ 3DEngine/3DEngine/App.cpp
@@ -82,7 +82,7 @@ void App::DoFrame()
 	{
 		if ( !wnd.CursorEnabled() )
 		{
-			cam.Rotate( delta->x,delta->y );
+			cam.Rotate( (float)delta->x,(float)delta->y );
 		}
 	}
